requirements.txt
Django~=3.2.10
django-crispy-forms~=1.14.0
requests~=2.28.1